Fix --enable-lint warnings.
Fix a few warnings thrown when compiling in lint-mode. No functional
or bugfix changes.

+++ fdisk_access.c
@@ -49,7 +49,7 @@ int partition_count(const char *devname)
         goto cleanup;
     }
 
-    result = fdisk_table_get_nents(table);
+    result = (int) fdisk_table_get_nents(table);
 
 cleanup:
     if (table) {
